make fmt

+++ kernel/src/syscall/mod.rs
@@ -977,8 +977,8 @@ impl Syscall {
 
             SYS_FCHOWN => {
                 let dirfd = args[0] as i32;
-                let uid = args[1] as usize;
-                let gid = args[2] as usize;
+                let uid = args[1];
+                let gid = args[2];
                 Self::fchown(dirfd, uid, gid)
             }
 
@@ -1142,21 +1142,21 @@ impl Syscall {
             }
             SYS_CHOWN => {
                 let pathname = args[0] as *const u8;
-                let uid = args[1] as usize;
-                let gid = args[2] as usize;
+                let uid = args[1];
+                let gid = args[2];
                 Self::chown(pathname, uid, gid)
             }
             SYS_LCHOWN => {
                 let pathname = args[0] as *const u8;
-                let uid = args[1] as usize;
-                let gid = args[2] as usize;
+                let uid = args[1];
+                let gid = args[2];
                 Self::lchown(pathname, uid, gid)
             }
             SYS_FCHOWNAT => {
                 let dirfd = args[0] as i32;
                 let pathname = args[1] as *const u8;
-                let uid = args[2] as usize;
-                let gid = args[3] as usize;
+                let uid = args[2];
+                let gid = args[3];
                 let flag = args[4] as u32;
                 Self::fchownat(dirfd, pathname, uid, gid, flag)
             }
